succcess the 5.1.1

+++ CH5/Chap01_UserInfo/src/main/java/com/example/chap01_userinfo/UserDaoJdbc.java
@@ -43,7 +43,7 @@ public class UserDaoJdbc implements UserDao {
                     user.getId(),
                     user.getName(),
                     user.getPassword(),
-                    user.getLevel(),
+                    user.getLevel().intValue(),
                     user.getLogin(),
                     user.getRecommend());
